Finished Animations

+++ GameEngine2.0/Source/Include/Math/Matrix/Matrix4.h
@@ -73,7 +73,7 @@ namespace Math {
 
 		Matrix4 GetCofactor(int p, int q, int n);
 
-		int GetDeterminant(int n);
+		float GetDeterminant(int n);
 
 		Matrix4 TransposeMatrix();
